исправления согласно парадигме безопасного программирования, рефакторинг clang'ом, реализованы тесты для matrix и parallel_column_sum

+++ parallel_column_sum.h
@@ -4,10 +4,7 @@
 #include "matrix.h"
 
 double* find_column_sum_parallel(const Matrix* matrix);
-int sum_one_column(const Matrix *matrix,
-                    int left_column,
-                    int right_column,
-                    int number_of_rows,
-                    double *shared_memory);
+int sum_one_column(const Matrix* matrix, long left_column, long right_column,
+                   size_t number_of_rows, double* shared_memory);
 
 #endif  // TP_C_CPP_HOMEWORK__PARALLEL_COLUMN_SUM_H_
